+++ scripts/index.py
@@ -23,6 +23,9 @@ PLACE_RELPATHS = [
     "data/curated/cities-major.csv",
     "data/curated/hydronyms-major.csv",
     "data/curated/oronyms-major.csv",
+    "data/curated/municipalities.csv",
+    "data/curated/hodonyms.csv",
+    "data/curated/microtoponyms.csv",
     "data/curated/agencies-foiv.csv",
     "data/curated/agencies-other.csv",
 ]